Add files via upload
All R Files for thesis work. This contains data cleaning for joint dataset 1 (RiskFreeRate.R) and for the joint dataset 2 (MarketReturn.R). Logistic regressions are also included on each of those for the corresponding dataset. For edge and non-edge cases of the market risk, logistic regressions are run on the R scripts with those names.